requirements.txt
# Core dependencies
google-genai==1.65.0
python-dotenv==1.2.1
requests==2.32.5
socksio==1.0.0

# Testing
pytest==9.0.2
pytest-cov==7.0.0

# API
fastapi==0.115.12
pydantic==2.10.6
httpx==0.28.1
python-multipart==0.0.12

# Database
sqlalchemy>=2.0

# OSS
oss2==2.19.1
